requirements.txt
requests
pyyaml
ollama